refactor(core): inline redundant aliases and wrappers
Import canonical selectors and schemas directly, use platform URL and crypto APIs at call sites, and fold the one-use page boundary mapping into its caller. Remove the obsolete re-export and wrapper tests with no behavior change.
