--- src/features/plans/read-projection/client.ts
@@ -1,1 +1,0 @@
-export { derivePlanSummaryDisplayStatus } from '@/features/plans/read-projection/selectors';
