+++ src/features/plans/read-projection/service.ts
@@ -1,11 +1,9 @@
-import type {
-  ModuleDetailReadModel,
-  PlanDbClient,
-} from '@/features/plans/read-projection/types';
+import type { ModuleDetailReadModel } from '@/features/plans/read-projection/types';
 import type {
   PlanListPage,
   PlanListQuery,
 } from '@/features/plans/read-projection/types';
+import type { DbClient } from '@/lib/db/types';
 import type { PaginationOptions } from '@/shared/constants/pagination';
 import type {
   ClientGenerationAttempt,
@@ -48,7 +46,7 @@ import { logger } from '@/lib/logging/logger';
 
 async function listPlanSummaries(params: {
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
   options?: PaginationOptions & { orderBy?: 'createdAt' | 'updatedAt' };
 }): Promise<PlanSummary[]> {
   const rows = await getPlanSummaryRowsForUser(
@@ -66,7 +64,7 @@ const DASHBOARD_PLAN_SUMMARY_LIMIT = 20 as const;
 
 export async function listDashboardPlanSummaries(params: {
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<PlanSummary[]> {
   return listPlanSummaries({
     userId: params.userId,
@@ -80,7 +78,7 @@ export async function listDashboardPlanSummaries(params: {
 
 export async function getPlansPageForRead(params: {
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
   query: PlanListQuery;
   referenceTimestamp?: string;
 }): Promise<PlanListPage> {
@@ -102,7 +100,7 @@ export async function getPlansPageForRead(params: {
 
 export async function listLightweightPlansForApi(params: {
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
   options?: PaginationOptions;
 }): Promise<LightweightPlanSummary[]> {
   const rows = await getLightweightPlanSummaryRowsForUser(
@@ -116,14 +114,14 @@ export async function listLightweightPlansForApi(params: {
 
 export async function listUsageAnalyticsPlanSummaries(params: {
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<LightweightPlanSummary[]> {
   return listLightweightPlansForApi(params);
 }
 
 export async function getPlanListTotalCount(params: {
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<number> {
   return getPlanSummaryCount(params.userId, params.dbClient);
 }
@@ -131,7 +129,7 @@ export async function getPlanListTotalCount(params: {
 export async function getPlanDetailForRead(params: {
   planId: string;
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<ClientPlanDetail | null> {
   const rows = await getLearningPlanDetailRows(
     params.planId,
@@ -164,7 +162,7 @@ export async function getPlanDetailForRead(params: {
 export async function getPlanGenerationStatusSnapshot(params: {
   planId: string;
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<PlanDetailStatusSnapshot | null> {
   const rows = await getPlanStatusRowsForUser(
     params.planId,
@@ -183,7 +181,7 @@ export async function getModuleLessonGenerationStatusForRead(params: {
   planId: string;
   moduleId: string;
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<{
   planId: string;
   moduleId: string;
@@ -210,7 +208,7 @@ export async function getModuleLessonGenerationStatusForRead(params: {
 export async function getPlanGenerationAttemptsForRead(params: {
   planId: string;
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<ClientGenerationAttempt[] | null> {
   const attempts = await getPlanAttemptsForUser(
     params.planId,
@@ -229,7 +227,7 @@ export async function getModuleDetailForRead(params: {
   planId: string;
   moduleId: string;
   userId: string;
-  dbClient?: PlanDbClient;
+  dbClient?: DbClient;
 }): Promise<ModuleDetailReadModel | null> {
   const rows = await getModuleDetailRows(
     params.planId,
