+++ src/features/plans/read-projection/types.ts
@@ -1,8 +1,5 @@
-import type { DbClient } from '@/lib/db/types';
 import type { ProgressStatus, ResourceType } from '@/shared/types/db.types';
 import type { LessonContent } from '@/shared/types/lesson-content.types';
-
-export type PlanDbClient = DbClient;
 
 export type PlanReadStatus =
   | 'not_started'
