--- src/features/plans/validation/shared.ts
@@ -1,10 +1,0 @@
-export {
-  NOTES_MAX_LENGTH,
-  TOPIC_MAX_LENGTH,
-} from '@/shared/constants/learning-plans';
-
-export {
-  LEARNING_STYLE_ENUM,
-  SKILL_LEVEL_ENUM,
-  weeklyHoursSchema,
-} from '@/shared/schemas/plan-validation.schemas';
