--- src/lib/api/coerce-unknown-to-message.ts
@@ -1,4 +1,0 @@
-/**
- * Best-effort string for thrown or unknown values (logging / API messages).
- */
-export { coerceUnknownToMessage } from '@/lib/errors/normalize-unknown';
