--- src/lib/api/page-boundary-result.ts
@@ -1,16 +1,0 @@
-/**
- * Maps a nullable request-boundary result to a typed access result with UNAUTHORIZED fallback.
- */
-export function finalizePageBoundaryResult<TSuccess, TUnauth>(
-  boundaryResult: TSuccess | null,
-  options: {
-    unauthenticatedMessage: string;
-    unauthenticated: (message: string) => TUnauth;
-  },
-): TSuccess | TUnauth {
-  if (boundaryResult !== null) {
-    return boundaryResult;
-  }
-
-  return options.unauthenticated(options.unauthenticatedMessage);
-}
